+++ objcToxTests/OCTSubmanagerBootstrapTests.m
@@ -105,8 +105,8 @@
 
     [self performBlock:^{
         [expectation fulfill];
-    } afterDelay:0.3];
-    [self waitForExpectationsWithTimeout:0.35 handler:nil];
+    } afterDelay:0.5];
+    [self waitForExpectationsWithTimeout:0.6 handler:nil];
 
     OCMVerify([self.tox bootstrapFromHost:@"h1" port:1 publicKey:@"1" error:[OCMArg anyObjectRef]]);
     OCMVerify([self.tox bootstrapFromHost:@"h2" port:2 publicKey:@"2" error:[OCMArg anyObjectRef]]);